pyproject.toml
[project]
name = "assist-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.13",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai>=0.3.27",
    "openai>=1.91.0",
    "posthog<6.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "torch>=2.2.2",
    "uuid>=1.30",
]
